Split PlayerStates implementations into separate files
